Renamed Avatar to NativeAvatar, JSAvatar to Avatar
